+++ superwall/src/main/java/com/superwall/sdk/paywall/view/webview/templating/models/DeviceTemplate.kt
@@ -29,6 +29,8 @@ data class DeviceTemplate(
     val timezoneOffset: Int,
     val radioType: String,
     val interfaceStyle: String,
+    val fontSize: Int,
+    val fontScale: Float,
     val isLowPowerModeEnabled: Boolean,
     val bundleId: String,
     val appInstallDate: String,